@SuppressWarnings("all")

+++ wukong-security/src/main/java/com.wukong.security/dao/UserDynamicSqlSupport.java
@@ -10,7 +10,7 @@ public final class UserDynamicSqlSupport {
     public static final User user = new User();
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
-    public static final SqlColumn<Long> userId = user.userId;
+    public static final SqlColumn<Integer> userId = user.userId;
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
     public static final SqlColumn<String> username = user.username;
@@ -29,7 +29,7 @@ public final class UserDynamicSqlSupport {
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
     public static final class User extends SqlTable {
-        public final SqlColumn<Long> userId = column("user_id", JDBCType.BIGINT);
+        public final SqlColumn<Integer> userId = column("user_id", JDBCType.INTEGER);
 
         public final SqlColumn<String> username = column("username", JDBCType.VARCHAR);
